refactor: Change placement for all utils finctions.
